Fix signed / unsigned comparsion warning
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@4053 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/oovother.cpp
@@ -4851,7 +4851,7 @@ RexxMethod2(logical_t, lv_setColumnOrder, RexxArrayObject, order, OSELF, self)
             RexxObjectPtr item;
             int column;
 
-            for ( int i = 0; i < items; i++)
+            for ( size_t i = 0; i < items; i++)
             {
                 item = context->ArrayAt(order, i + 1);
                 if ( item == NULLOBJECT || ! context->ObjectToInt32(item, &column) )
